src/main/java/ca/wendyliu/spring5recipeapp/service/UnitOfMeasureServiceImpl.java
package ca.wendyliu.spring5recipeapp.service;

import ca.wendyliu.spring5recipeapp.commands.UnitOfMeasureCommand;
import ca.wendyliu.spring5recipeapp.converters.UnitOfMeasureToUnitOfMeasureCommand;
import ca.wendyliu.spring5recipeapp.domain.UnitOfMeasure;
import ca.wendyliu.spring5recipeapp.repository.UnitOfMeasureRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Set;
import java.util.stream.Collectors;
import java.util.stream.StreamSupport;

@Service
public class UnitOfMeasureServiceImpl implements UnitOfMeasureService {

    private final UnitOfMeasureRepository unitOfMeasureRepository;
    private final UnitOfMeasureToUnitOfMeasureCommand converter;

    @Autowired
    public UnitOfMeasureServiceImpl(UnitOfMeasureRepository unitOfMeasureRepository,
                                    UnitOfMeasureToUnitOfMeasureCommand converter) {
        this.unitOfMeasureRepository = unitOfMeasureRepository;
        this.converter = converter;
    }

    @Override
    public Set<UnitOfMeasureCommand> listAllUoms() {
        return StreamSupport.stream(unitOfMeasureRepository.findAll()
                .spliterator(), false)
                .map(converter::convert)
                .collect(Collectors.toSet());
    }
}
